requirements.txt
flask==2.3.3
sentence-transformers==2.2.2
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.35.2
tokenizers==0.15.0
huggingface-hub==0.19.4
pymongo==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0 